requirements.txt
certifi==2017.4.17
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
numpy==1.13.1
python-firebase==1.2
requests==2.18.1
six==1.10.0
urllib3==1.21.1
Werkzeug==0.12.2
